Scaffold Extension (#3)
* Scaffold Extension

This scaffolds out a HCL VS Code extension. It borrows heavily from the existing vscode-terraform setup for configuration files and GitHub Actions.

A minimal hcl.tmGrammar.json is committed along with the extension, but is intended to be replaced in the future by the content in the hashicorp/syntax repo.