Walk back Language activationEvents
This removes the `.nomad` file extension from activationEvents

Previously the intention was to handle every HashiCorp product file extension that uses HCL as it's language, and then remove from this extension if/when new VS Code extensions are made. We're going to take the pragmatic approach of providing HCL only first and then see how user adoption flows.