Update README content (#73)
This adds more detail to the README and works on standardizing naming.

This changed all references to the extension as `HCL Extension for Visual Studio Code` to comply with https://code.visualstudio.com/brand. It also added boilerplate info from the HCL repo to explain what it is we are actually highlighting.