Enable VS Code Web Extension support (#114)
This is primarily a syntax extension, so it should work in VS Code web environments like github.dev and vscode.dev. Since we have actual code we show up as a normal extension and not one that can work on the web.

This commit adds a web extension file, supporting metadata to the package manifest, and ports bundling the extension to webpack. The result is a single VSIX that supports both desktop and web deployments.

In addition, this also updates all the VS Code task and launch configurations so that the extension can be tested locally both as a desktop or web extension.

**Note** that there is now a `src/extension.ts` and a `src/web/extension.ts` that appear identical. The two locations are required along with the changes inside package.json in order for the extension to be bundled properly for the two different environments. Most importantly, they allow webpack to properly include either the browser or desktop version of @vscode/extension-telemetry in the approprite extension. Consult https://code.visualstudio.com/api/extension-guides/web-extensions#web-extension-anatomy for more information about these requirements.