Measure usage of known products
Having an idea of which products are used by the user base would be useful for future planning. Such data could back up any potential plans and decisions around dedicated extensions for individual products. For example, if we knew that e.g. 60% users open Packer files and 30% Waypoint, then that gives us indication on what extensions are in demand and that could inform future plans.